Update feature location NLP + Preferences Really slow need to be fixed

+++ org.but4reuse.wordclouds/src/org/but4reuse/wordclouds/similarity/PreferenceInitializer.java
@@ -6,16 +6,14 @@ import org.eclipse.jface.preference.IPreferenceStore;
 
 public class PreferenceInitializer extends AbstractPreferenceInitializer {
 
-
 	@Override
 	public void initializeDefaultPreferences() {
 
 		IPreferenceStore store = Activator.getDefault().getPreferenceStore();
-		
+
 		store.setDefault(TermBasedConstraintsDiscoveryPreferences.EXCLUDES, true);
 		store.setDefault(TermBasedConstraintsDiscoveryPreferences.REQUIRES, true);
-		store.setDefault(TermBasedConstraintsDiscoveryPreferences.HIGHT_SIM, (double)0.5);
-		
+		store.setDefault(TermBasedConstraintsDiscoveryPreferences.HIGHT_SIM, 0.5);
 
 	}
