+++ org.but4reuse.wordclouds/src/org/but4reuse/wordclouds/util/WordCloudUtil.java
@@ -161,10 +161,10 @@ public class WordCloudUtil {
 
 		}
 
-		if(cpt == 0)
+		if (cpt == 0)
 			return 0;
-		
-		return (double) (res) / cpt;
+
+		return (res) / cpt;
 	}
 
 	/**
